课程接口

+++ src/main/java/com/scut/GymManager/service/impl/CourseInfoServiceImpl.java
@@ -1,9 +1,11 @@
-﻿package com.scut.GymManager.service.impl;
+package com.scut.GymManager.service.impl;
 
 import java.util.List;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 
+import com.scut.GymManager.utility.JwtUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,12 +23,31 @@ import com.scut.GymManager.service.CourseInfoService;
 public class CourseInfoServiceImpl implements CourseInfoService {
 
 	private Logger log=LoggerFactory.getLogger(getClass());
+
 	@Resource
 	private CourseInfoMapper courseInfoMapper;
+
+	@Resource
+	private JwtUtil jwtUtil;
+
+	@Resource
+	private HttpServletRequest httpServletRequest;
+
 	@Override
 	public void createCourse(CourseRequest courseRequest) throws CrudException {
-		CourseInfo courseInfo=CourseInfo.builder().CourseName(courseRequest.getCourseName()).CourseTime(courseRequest.getCourseTimes())
-				.MaxNumber(courseRequest.getMaxNumber()).Classroom(courseRequest.getClassroom()).coachId(courseRequest.getCoachId()).studentNum(courseRequest.getStudentNum())
+
+		String identity = jwtUtil.extractIdentitySubject(this.httpServletRequest);
+
+		if (!identity.equals("Coach"))
+			throw new CrudException("非教练不能创建课程");
+
+		CourseInfo courseInfo=CourseInfo.builder()
+				.CourseName(courseRequest.getCourseName())
+				.CourseTime(courseRequest.getCourseTime())
+				.MaxNumber(courseRequest.getMaxNumber())
+				.Classroom(courseRequest.getClassroom())
+				.coachId(courseRequest.getCoachId())
+				.studentNum(courseRequest.getStudentNum())
 				.build();
 		int x=courseInfoMapper.insert(courseInfo);
 		if(x==0) throw new CrudException("insert 出错");
@@ -34,12 +55,24 @@ public class CourseInfoServiceImpl implements CourseInfoService {
 
 	@Override
 	public void updateCourse(CourseInfo courseInfo) throws CrudException {
+
+		String uid = jwtUtil.extractUidSubject(this.httpServletRequest);
+
+		if (!courseInfo.getCoachId().equals(uid))
+			throw new CrudException("你没有该操作权限");
+
 		int x=courseInfoMapper.updateById(courseInfo);
 		if(x==0) throw new CrudException("update 出错");
 	}
 
 	@Override
 	public void deleteCourse(String CourseId) throws CrudException {
+
+		String uid = jwtUtil.extractUidSubject(this.httpServletRequest);
+
+		if (!uid.equals("1"))
+			throw new CrudException("你没有操作权限");
+
 		int x=courseInfoMapper.deleteById(CourseId);
 		if(x==0) throw new CrudException("delete 出错");
 	}
@@ -52,13 +85,13 @@ public class CourseInfoServiceImpl implements CourseInfoService {
 
 	@Override
 	public List<CourseInfo> viewCoachCourseTable(String coachId) {
-		
+
 		return courseInfoMapper.searchCoachList(coachId);
 	}
 
 	@Override
-	public List<CourseInfo> viewVIPCourseTable(String VIPId) {
-		return courseInfoMapper.searchVIPList(VIPId);
+	public List<CourseInfo> viewVIPCourseTable(String vipId) {
+		return courseInfoMapper.searchVIPList(vipId);
 	}
 
 }
